moving keygen into thelib
